requirements.txt
Django==4.2.5
django-admin-sortable2==2.1.9
